requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
bootstrap4==0.1.0
colorama==0.4.4
Django==3.1.4
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-markdown-deux==1.0.5
django-pagedown==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Dumper==1.2.0
factory-boy==3.2.0
Faker==5.3.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
markdown2==2.3.10
mccabe==0.6.1
Pillow==8.0.1
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
